backend/user_service/api/requirements.txt
fastapi
uvicorn
SQLAlchemy
python-dotenv
psycopg2
aio_pika
pika
asyncio